requirements.txt
numpy==1.11.1rc1
gym==0.10.8
gym-retro==0.6.0
